Move payload annotation to annotation package.
