+++ src/main/java/ysoserial/payloads/C3P0.java
@@ -15,8 +15,8 @@ import javax.sql.PooledConnection;
 import com.mchange.v2.c3p0.PoolBackedDataSource;
 import com.mchange.v2.c3p0.impl.PoolBackedDataSourceBase;
 
-import ysoserial.PayloadTest;
 import ysoserial.payloads.annotation.Dependencies;
+import ysoserial.payloads.annotation.PayloadTest;
 import ysoserial.payloads.util.PayloadRunner;
 import ysoserial.payloads.util.Reflections;
