+++ src/main/java/ysoserial/payloads/annotation/PayloadTest.java
@@ -1,4 +1,4 @@
-package ysoserial;
+package ysoserial.payloads.annotation;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
