+++ src/test/java/ysoserial/payloads/PayloadsTest.java
@@ -26,13 +26,13 @@ import org.junit.runners.Parameterized.Parameters;
 import ysoserial.CustomPayloadArgs;
 import ysoserial.CustomTest;
 import ysoserial.Deserializer;
-import ysoserial.PayloadTest;
 import ysoserial.Serializer;
 import ysoserial.Throwables;
 import ysoserial.WrappedTest;
 import ysoserial.payloads.TestHarnessTest.ExecMockPayload;
 import ysoserial.payloads.TestHarnessTest.NoopMockPayload;
 import ysoserial.payloads.annotation.Dependencies;
+import ysoserial.payloads.annotation.PayloadTest;
 import ysoserial.payloads.util.ClassFiles;
 import ysoserial.secmgr.ExecCheckingSecurityManager;
 import ysoserial.secmgr.ExecCheckingSecurityManager.ExecException;
